Fixing an issue with the UglifyJS' mangler
UglifyJS's mangler used to change renderer's function names to simple characters so the whole Renderer's cache thing (used by renderers or plugins) wasn't working.
So, we dropped using function.name (which was bad in the first place) and added a static property "id" for every renderers (a small drawback for developers).